% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{build_cruise_info}
\alias{build_cruise_info}
\title{Build a descriptive cruise info string from sample timestamps}
\usage{
build_cruise_info(sample_times)
}
\arguments{
\item{sample_times}{POSIXct vector of sample timestamps.}
}
\value{
Character string.
}
\description{
Determines the most common month among samples and formats as
"RV Svea \if{html}{\out{<Month>}} cruise, \if{html}{\out{<start date>}} to \if{html}{\out{<end date>}}".
}
\keyword{internal}
