% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_cache.R
\name{metadata_cache_path}
\alias{metadata_cache_path}
\title{Path of the metadata cache file inside local storage}
\usage{
metadata_cache_path(storage_dir)
}
\arguments{
\item{storage_dir}{Local storage base directory.}
}
\value{
Character path.
}
\description{
Path of the metadata cache file inside local storage
}
\keyword{internal}
